+++ Models/Base/IEntityBaseRepository.cs
@@ -7,11 +7,11 @@ namespace Director.Models.Base
 {
    public interface IEntityBaseRepository<T> where T:class, IEntityBase, new()
     {
-        Task<IEnumerable<T>> GetListByIdAsync(long id);
+        Task<IEnumerable<T>> GetListByIdAsync(int id);
         Task<IEnumerable<T>> GetAllAsync();
-        Task<T> GetByIdAsync(long id);
+        Task<T> GetByIdAsync(int id);
         Task AddAsync(T entity);
-        Task UpdateAsync(long id, T entity);
-        Task DeleteAsync(long id);
+        Task UpdateAsync(int id, T entity);
+        Task DeleteAsync(int id);
     }
 }
